chore: rename to useAllowedSlippage
